Refactoring app
